Introducing the new SMS Aggregator !
Available on /aggregator/sms, support post and get.
Full Oriented Object, And Full Exception Raised during the process

An SMS is composed by a from (the phone number of the client), a to (the short of balloon) and a content (ie: HEC 1, or HEC What's the point)
Here is the list of the exception that can be raised :

EventNotFoundException
NoSurveyAvailableException
TooManySurveyAvailableException
TooManyWallFoundException
UserHasAlreadyVoteOnThisSurveyException
WallNotFoundException.

Try it, love it ;)